requirements.txt
# Core dependencies
openai>=1.30.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
aiosqlite>=0.20.0

# Testing and development
unittest-xml-reporting>=3.2.0

# Browser MCP integration notes:
# - Browser MCP server installed via npm: npm install -g @browsermcp/mcp@latest
# - Direct JSON-RPC communication (no Python MCP SDK dependency needed)
# - Requires Node.js and Chrome with Browser MCP extension


